[FIRRTL] Implement LiteralDataTapKey support
Add support for `LiteralDataTapKey`s to the `GrandCentralTaps` pass.
The change to GCT itself is simple enough, merely adding additional
connects that drive the constant literal to the appropriate output port.

This is the first time as far as I can tell that we have the need to
call the FIR parser on some string buffer. This commit adds a dedicated
`firrtl::parseIntegerLiteralExp` that takes an input `StringRef`, sets
up the entire parser machinery, and then treats the input string as an
integer literal.

Annotations have been spotted in the wild which contain snippets of
FIRRTL as strings, and probably need parsing/inlining at some point.
Those annos will benefit from figuring the parsing out.

Fixes #1384.