[LowerToHW] Allow sv::InterfaceOps to pass through
Add support for operations of specific types to pass through FIRRTL to
HW lowering.  This uses a vector of operations which are inserted into
the outer module after FIRRTL modules are lowered.

Use this to allow SystemVerilog InterfaceOps to pass through.
Previously, these would cause an error and get dropped.

This is done to allow for FIRRTL transforms to add operations which may
have no representation in the FIRRTL dialect.  This is viewed as better
than, alternatively, encoding operations in annotations/attributes and
having LowerToHW expand them.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>